Move original spreadsheet factory to juancarlos one
