+++ src/main/java/edu/upc/etsetb/arqsoft/miniexceljc/util/S2VSpreadsheetLoader.java
@@ -1,12 +1,12 @@
-package edu.upc.etsetb.arqsoft.miniexceljc.model;
+package edu.upc.etsetb.arqsoft.miniexceljc.util;
 
 import edu.upc.etsetb.arqsoft.miniexceljc.client.text.InvalidInputException;
+import edu.upc.etsetb.arqsoft.miniexceljc.factories.SpreadsheetFactory;
+import edu.upc.etsetb.arqsoft.miniexceljc.model.ContentSpec;
+import edu.upc.etsetb.arqsoft.miniexceljc.model.CoordinateSpec;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.Content;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.Coordinate;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.Spreadsheet;
-import edu.upc.etsetb.arqsoft.miniexceljc.util.AlphabeticRadixConverter;
-import edu.upc.etsetb.arqsoft.miniexceljc.util.NumericalContentChecker;
-import edu.upc.etsetb.arqsoft.miniexceljc.util.TextContentChecker;
 
 import java.io.File;
 import java.io.FileNotFoundException;
