+++ src/main/java/edu/upc/etsetb/arqsoft/miniexceljc/util/SpreadsheetSaver.java
@@ -1,5 +1,6 @@
-package edu.upc.etsetb.arqsoft.miniexceljc.model;
+package edu.upc.etsetb.arqsoft.miniexceljc.util;
 
+import edu.upc.etsetb.arqsoft.miniexceljc.factories.SpreadsheetFactory;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.Spreadsheet;
 
 import java.util.logging.Logger;
