+++ src/main/java/edu/upc/etsetb/arqsoft/miniexceljc/view/text/TextUISpreadsheet.java
@@ -1,7 +1,7 @@
 package edu.upc.etsetb.arqsoft.miniexceljc.view.text;
 
+import edu.upc.etsetb.arqsoft.miniexceljc.factories.SpreadsheetFactory;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.CoordinateSpec;
-import edu.upc.etsetb.arqsoft.miniexceljc.model.SpreadsheetFactory;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.Coordinate;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.Value;
 import edu.upc.etsetb.arqsoft.miniexceljc.util.AlphabeticRadixConverter;
